Conflicts:
	src/main/java/com/supinfo/supapi/entity/User.java

+++ src/main/java/com/supinfo/supapi/entity/Line.java
@@ -15,9 +15,10 @@ import com.supinfo.supapi.entity.association.StationLineAssociation;
 
 
 @Entity
+//@JsonIdentityInfo(generator = ObjectIdGenerators.IntSequenceGenerator.class, property = "id")
 public class Line {
 	@Id
-	private int id;
+	private long id;
 	
 	private String reference;
 	private String name;
